271191b130efc38ca9da51d1f44a796bf3a95dab Release 7.0.0 (Dr Maxim Orlovsky)

Pull request description:

  Rename SigHash* into Sighash accross the repo

  Reproducible with automatic case-sensetive search/replace

  Make taproot leaf script depth to be u8 everywhere

  I have changed the release notes title from 7.0.0-rc.2 to 7.0.0.

ACKs for top commit:
  apoelstra:
    ACK 271191b130efc38ca9da51d1f44a796bf3a95dab

Tree-SHA512: c77bddaef4ac0862be9920c1d65a6abf2e67226b3b18ce9f873ddfa33d2889427fb05f29cf31174b00fe2849ee84ddb006f8e8632b3ba4497abaf42e66d1fe75

+++ src/descriptor/tr.rs
@@ -313,14 +313,14 @@ pub struct TapTreeIter<'a, Pk: MiniscriptKey>
 where
     Pk: 'a,
 {
-    stack: Vec<(usize, &'a TapTree<Pk>)>,
+    stack: Vec<(u8, &'a TapTree<Pk>)>,
 }
 
 impl<'a, Pk> Iterator for TapTreeIter<'a, Pk>
 where
     Pk: MiniscriptKey + 'a,
 {
-    type Item = (usize, &'a Miniscript<Pk, Tap>);
+    type Item = (u8, &'a Miniscript<Pk, Tap>);
 
     fn next(&mut self) -> Option<Self::Item> {
         while !self.stack.is_empty() {
@@ -681,8 +681,8 @@ impl<P: MiniscriptKey, Q: MiniscriptKey> TranslatePk<P, Q> for Tr<P> {
 }
 
 // Helper function to compute the len of control block at a given depth
-fn control_block_len(depth: usize) -> usize {
-    TAPROOT_CONTROL_BASE_SIZE + depth * TAPROOT_CONTROL_NODE_SIZE
+fn control_block_len(depth: u8) -> usize {
+    TAPROOT_CONTROL_BASE_SIZE + (depth as usize) * TAPROOT_CONTROL_NODE_SIZE
 }
 
 // Helper function to get a script spend satisfaction
